backend_ml/imgtotxt/requirements.txt
Flask
PyMuPDF
easyocr
google-generativeai
werkzeug
numpy